--- go/testutils/doc.go
@@ -1,2 +1,0 @@
-// Package testutils provides golang testutils for developing with TemporalX
-package testutils
