+++ go/xtestutils/doc.go
@@ -1,0 +1,2 @@
+// Package xtestutils provides golang testutils for developing with TemporalX
+package xtestutils
