+++ go/xtestutils/testutils.go
@@ -1,4 +1,4 @@
-package testutils
+package xtestutils
 
 import (
 	"os"
