+++ go/xtestutils/testutils_test.go
@@ -1,4 +1,4 @@
-package testutils
+package xtestutils
 
 import (
 	"os"
